requirements.txt
jinja2
networkx
protobuf
PyYAML
requests
urllib3